+++ src/pages/home/components/product-list/ProductList.jsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { getProductList } from "@/redux/actions";
 
 import Loading from "@/pages/loading/Loading";
-import Error from "@/pages/error/components/Error";
+import Error from "@/pages/error/Error";
 
 import StyledProductList from "@/pages/home/components/product-list/ProductList.styled";
 import ProductItem from "@/pages/home/components/product-item/ProductItem";
